refactor: perceptortab
